added pfinderSim
pfinderSim is simulates the Lost Princess game in python. from a given
location, a set of valid moves is constructed (the game limits the tiles
that you can visit from your current location to a rectangular window).
travelling to a tile has a cost of 2 + 0.9*distance, and searching a
tile has a cost of 5, reflecting the time to perform that action in
game. running the search will initialize the princess in a random
location (with the ability to set a seed) and the player at a
location(either given, random, or random seeded), and runs a given
policy to search (interactive policy currently implemented).